+++ nsjail.h
@@ -157,7 +157,7 @@ struct nsjconf_t {
 	size_t cgroup_mem_max;
 	const char* cgroup_pids_mount;
 	const char* cgroup_pids_parent;
-	size_t cgroup_pids_max;
+	unsigned int cgroup_pids_max;
 	const char* cgroup_net_cls_mount;
 	const char* cgroup_net_cls_parent;
 	unsigned int cgroup_net_cls_classid;
